% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR.R
\name{add_noise}
\alias{add_noise}
\title{Complement the missing variances of reconstructed variables by appropriate amount of independent white noise}
\usage{
add_noise(Xg, X)
}
\arguments{
\item{Xg}{l}

\item{X}{l}
}
\value{
An updated matrix \code{Xg} with random noise added to each of the columns.
}
\description{
Add \code{E_i} of noise from normal distribution to each reconstructed variable \code{Xg_i} such that \code{var(Xg_i)==var(X_i)} where \code{X_i} is a reference variable.
}
