% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR.R
\name{allocate_blocks}
\alias{allocate_blocks}
\title{Place blockwise PCA reconstructions in an input container matrix}
\usage{
allocate_blocks(where, blockwise_PCA)
}
\arguments{
\item{where}{Input matrix where rows are samples and columns variables. It should have the same dimensions as matrix used as input to \code{blockwise_PCA_reconstruction} function used to produce \code{blockwise_PCA} argument.}

\item{blockwise_PCA}{Result of \code{blockwise_PCA_reconstruction} function ot use.}
}
\value{
An updated matrix \code{where} with columns corresponding to each cluster were replaced by their recondstructions from \code{blockwise_PCA} objects.
}
\description{
Place blockwise PCA reconstructions in an input container matrix
}
\examples{
data(brca)
data(brca_clusters)
cl_list<- clustering_vector2list( brca_clusters ) 
blocks_brca<- blockwise_PCA_reconstruction(A=brca, A_variances= matrixStats::colVars(brca), clustering_list= cl_list, f=0.6, k=3)
brca_rec<- brca
brca_rec[,]=0
brca_rec= allocate_blocks( brca_rec, blocks_brca)
}
\seealso{
\code{\link[=blockwise_PCA_reconstruction]{blockwise_PCA_reconstruction()}}
}
