% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR.R
\name{blockwise_PCA_reconstruction}
\alias{blockwise_PCA_reconstruction}
\title{Perform k-th order PCA reconstruction of correlated data with clustered variables}
\usage{
blockwise_PCA_reconstruction(A, A_variances, clustering_list, f, k)
}
\arguments{
\item{A}{Data matrix where columns are variables and rows are samples.}

\item{A_variances}{Precomputed variances of columns of \code{A}. If not computed already, set this to \code{colVars(A)}}

\item{clustering_list}{A named list where names are cluster labels (integers converted to strings) and elemnts are vectors of indexes of nodes (columns of \code{A}) in each cluster, as produced by \code{clustering_vector2list} or outlined in \code{HCCSim_clustering_list}.}

\item{f}{Target fraction of variance to explain, a real number in \code{(0,1]}.}

\item{k}{Maximum number of PCs to use per cluster in \code{clustering_list}.}
}
\value{
A \code{list} with one element per cluster in \code{clustering_list}. Each such element contains the reconstruction of columns of \code{A} in that cluster and additional information on the result, i.e. each element is a list with the following elements:
\itemize{
\item \code{location} A vector of indexes of reconstructed columns in the block as placed in \code{A}.
\item \code{total_var} Real number, total variance of the block.
\item \code{pc_var} Variances of all PCs of each block.
\item \code{cumul_var} Vector of cumulative variances of all PCs.
\item \code{k_used} Number of PCs used (can be less than \code{k} )
\item \code{k_f} Number of PCs one would have to use to explain target fraction \code{f} of total variance.
\item \code{f} Same as input argument.
\item \code{saturated} True if total variance was fully explained with \code{k_used} PCs.
\item \code{var_explained} How much variance in total was explained.
\item \code{PC} Matrix of PCs used for reoconstruction, has \code{k2use} columns, one per each PC.
\item \code{Vt} Coefficients of PCs used for reconstruction.
\item \code{reconstruction} Reconstructed data using \code{PC} and \code{Vt}.
}
}
\description{
In a nutshell, function performs PCA separately in each group of columns of \code{A} defined by \code{clustering_list} using at most \code{k} PCs per group of columns, trying to explain a fraction \code{f} of total variance inside each group and uses the result to reconstruct \code{A} using those clusterwise PCs.
}
\examples{
data(brca)
data(brca_clusters)
cl_list<- clustering_vector2list( brca_clusters ) 
blocks_brca<- blockwise_PCA_reconstruction(A=brca, A_variances= matrixStats::colVars(brca), clustering_list= cl_list, f=0.6, k=3)
}
