% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRCA_GE-data.R
\docType{data}
\name{brca}
\alias{brca}
\title{BRCA gene expression dataset}
\format{
A large numeric matrix where columns are genes and rows are samples.
}
\usage{
data(brca)
}
\description{
Data description
}
\examples{
data(brca)
dim(brca)
}
\references{
Add some
}
\keyword{datasets}
