% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR.R
\name{calculate_f}
\alias{calculate_f}
\title{Recalculate fraction of variance to explain taking the clusterless variables into the account}
\usage{
calculate_f(clustering_vector, f_E, X_variances)
}
\arguments{
\item{clustering_vector}{An integer vector for which \code{clustering_vector[i]} gives a label of the group to which node \code{i} belongs. Label of \code{0} should encode clusterless nodes.}

\item{f_E}{Target fraction of total variance in the dataset to explain.}

\item{X_variances}{Precalculated variances of each variable in the dataset, should be the same length and ordering as in \code{clustering_vector}. If not calculated already, set this to \code{colVars(X)} where \code{X} is matrix where rows are samples and columns variables.}
}
\value{
A number in \code{(0,1]}, recalulated \code{f_E} such that the recalculated fraction takes into the account only the variables in proper clusters (label \code{!=0} ).
}
\description{
Recalculate fraction of variance to explain taking the clusterless variables into the account
}
\examples{
data(brca)
data(brca_clusters)
calculate_f( zeroOutSmall(clustering_vector, 30), f_E=0.7, X_variances= colVars(brca) )
}
