% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{ccfast}
\alias{ccfast}
\title{Compute weighted version of clustering coefficient from WGCNA utilizing matrix product}
\usage{
ccfast(A)
}
\arguments{
\item{A}{symmetric matrix of similarities between nodes, \code{A[i,j]} is weight of the connection between nodes \code{i,j}. All entries must be between \code{0} and \code{1}.}
}
\value{
A vector of clustering coefficients - weighted version of the clustering coefficient designed for weighted correlation networks, as proposed by \code{WGCNA} methodology.
}
\description{
This function produces clustering coefficient from similarity matrix just like \code{clusterCoef} from \code{WGCNA} package. It is however (experimentally) much more time efficient because it explicitly involves a (matrix-matrix) matrix product. Speed up is drastic if R uses efficient matrix multiplication backend like OpenBLAS.
}
\examples{
data(brca)
S<- similarity_matrix(fastPearsonData(brca))
cc<- ccfast(S)
}
