% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCS.R
\name{cloned_normalPC_matrix}
\alias{cloned_normalPC_matrix}
\title{Get set generator PCs correlated like reference ones}
\usage{
cloned_normalPC_matrix(reference_PC_gen_matrix)
}
\arguments{
\item{reference_PC_gen_matrix}{A matrix of PCs of clusters and subclusters which correlations are to be replicated, such as one prouced by \code{PC_generator_matrix}}
}
\value{
A matrix of the same dimensions as input one, containing synthetic PCs from normal distribution with identical covariance as reference ones but independent of original ones.
}
\description{
Get set generator PCs correlated like reference ones
}
\examples{
data(brca)
data(brca_clusters)
lvl1<- initial_clusterNreconstruct(X= brca, X_variances=matrixStats::colVars(brca),
			        clustering_vector=brca_clusters)
lvl2<- subclusterNreconstruct(X=brca,
				 X_variances= matrixStats::colVars(brca),
				 hclustering_list= lvl1$clustering_list,
                              hcluster_PCA_blocks=lvl1$cluster_blockPCA,
                              clfun2=similarity_based_hclust,
                              clfun2OtherArgs_constant=list(method="complete" ),
			 clfun2OtherArgs_ranges= list(n_group=2:7)) 
PC_generator_matrix( lvl2$hcluster_blockPCA) -> PCmat
cloned_PC<- cloned_normalPC_matrix(PCmat)
 #note: no correlations between cloned and reference PCs
heatmap( abs(cor(cbind(PCmat, cloned_PC))), Colv=NA, Rowv=NA)
}
