% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_construction.R
\name{corfast}
\alias{corfast}
\title{Minimal optimized computation of Pearson correlation by matrix product}
\usage{
corfast(D)
}
\arguments{
\item{D}{Must be an object of a subclass of \code{matrix}. Rows store samples and columns - variables. No \code{NA}s allowed.}
}
\value{
A numeric square matrix of Pearson correlation coefficient compute between each pair of varables in \code{D}.
}
\description{
This function will generally perform faster than base R \code{cor}, provided one uses OpenBLAS matrix multiplication backend.
}
\examples{
data(brca)
C<- corfast(brca)
}
