% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_construction.R
\name{corpval}
\alias{corpval}
\title{Minimal optimized computation of a p-value of a Pearson correlation coefficient}
\usage{
corpval(r, n)
}
\arguments{
\item{r}{A vector of correlation coefficients for which to compute p-values.}

\item{n}{A number of samples used to compute each statistic in \code{r}. Assumed to be the same for all entries of \code{r}.}
}
\value{
A vector of corresponding 2-tailed p-values of statistics in \code{r}, where under the null hypothesis \code{r} assumes \code{T} distribution with \code{n-2} degrees of freedom.
}
\description{
This function will generally perform faster than base R \code{cor.test} but assumes much about the cleaniness of the data input.
}
\examples{
data(brca)
C<- corfast(brca)
pv<- corpval(C[lower.tri(C)], nrow(brca))
}
