% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{ecdf_distance}
\alias{ecdf_distance}
\title{Compute distance between empricial distribution functions}
\usage{
ecdf_distance(y, ecdf_x, summarizeFun = NULL)
}
\arguments{
\item{y}{A input vector of values which \code{ecdf} will be compared with the reference.}

\item{ecdf_x}{A function acting on \code{y}, reference empirical c.d.f. (or theoretical of a known distribution), e.g. produced by base R \code{ecdf()} function.}

\item{summarizeFun}{Optionally, an aggregation function called on the vector of absolute differences between two c.d.f.s.}
}
\value{
If \code{summarizeFun} is \code{NULL} (default), a vector of absoulute differences between two compared c.d.f's computed on input \code{y}. If \code{summarizeFun} is given, return value of \code{summarizeFun} called on the mentioned vector of differences  is returned instead.
}
\description{
Given and input value vector \code{y} and prepared \code{ecdf} object (empirical distribution function of other distribution), function will compute distance between empirical c.d.f. of \code{y} and the input \code{ecdf}.
}
\examples{
x<- rnorm(1000)
y<- rnorm(1000,0.05)
diffs<- ecdf_distance(x, ecdf(y))
mean_dist<- ecdf_distance(x, ecdf(y), mean)
}
\seealso{
\code{\link[=KS_ecdf_distance]{KS_ecdf_distance()}} for specific version of this function
}
