% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_construction.R
\name{fastPearsonData}
\alias{fastPearsonData}
\title{Generate list storing all pairwise Pearson correlation coefficients between all variables in the dataset, and their (adjusted) 2-tailed p-values using \code{corfast}}
\usage{
fastPearsonData(D, NA.warn = TRUE, p.adjust.method = "holm")
}
\arguments{
\item{D}{Must be an object of a subclass of \code{matrix}. Rows store samples and columns - variables. No checks on 0 variance are performed.}

\item{NA.warn}{Logical, if true (default), then a warning will be displayed if \code{D} contains any \code{NA} values.}

\item{p.adjust.method}{The method of p-value adjustment used. For \code{D} with \code{N} columns, number of tests is \code{N*(N-1)*0.5}. Gets passed to argument \code{p.adjust.method} in a call of \code{p.adjust} function.}
}
\value{
A list storing all the important data about correlations in \code{D} with the following fields:
\itemize{
\item \code{N} Number of variables in \code{D}.
\item \code{r} A vector of all correlation coefficients between all pairs of variables in \code{D}.
\item \code{P} A vector of unadjusted p-values of \code{r}.
\item \code{Pa} A vector of adjusted p-values of \code{r}.
}
}
\description{
Function will run very fast if efficient matrix multiplication backed is used (like OpenBLAS). However, no input correctness checks are performed (apart from warning on presence of \code{NA} in input).
}
\examples{
data(brca)
fastPearsonData(brca) -> brca_corData
C<- matrix(nrow= brca_corData$N, ncol= brca_corData$N) # restore matrices from vectors in corData like this:
C[lower.tri(C)]= brca_corData$r
C<-t(C)
C[lower.tri(C)]= brca_corData$r
diag(C)=1
}
