% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR.R
\name{get_full_reconstruction}
\alias{get_full_reconstruction}
\title{Perform the full reconstruction from HCD of a dataset}
\usage{
get_full_reconstruction(
  X,
  hcluster_blockPCA,
  hclustering_list,
  add_noise = TRUE,
  uncenter = TRUE
)
}
\arguments{
\item{X}{l}

\item{hcluster_blockPCA}{l}

\item{hclustering_list}{l}

\item{add_noise}{l}

\item{uncenter}{l}
}
\value{
A matrix \code{Xg} which is the \code{g}-th order hierarchical clustering reconstruction of correlated variables in \code{X}, where \code{g} is the number of levels in \code{hclustering_list} and \code{hcluster_blockPCA}.
}
\description{
Given a HCD (hierarchical clustering decomposion of correlation structure) of a dataset, generate it's reconstruction (HCR) based on that HCD.
}
