% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR.R
\name{get_max_lvl}
\alias{get_max_lvl}
\title{Extract number of levels of hierarchy from \code{hclustering_list}}
\usage{
get_max_lvl(hclustering_list)
}
\arguments{
\item{hclustering_list}{A list of clusters and sub-clusters indexed by strings of integer tuples (like '1','2','1,1','1,2','2,1','2,2' etc.). Each cluster is a vector of integers dentoing nodes belonging to that cluster. Designed to work with output of \code{subdivide_cluster} function or instance of \code{HCCsim_clustering_list} class.}
}
\value{
An integer giving the number of levels of hierarchy in \code{hclustering_list}
}
\description{
Extract number of levels of hierarchy from \code{hclustering_list}
}
\examples{
base_cl<- HCCSim_clustering_list(list( `1`= c(5,6,7,8), `2`= c(9,4,3), `3`=c(1,2,10,11) ),
				    domain_size= 15,   #4 clusterless nodes implied
				    params=list(base_param=1) )
subdivision1= HCCSim_clustering_list(list(c(6,8), c(7,5) ), params=list(sub1_param=2))
subdivision2= HCCSim_clustering_list(list(c(1,10), c(2,11) ), params=list(sub2_param=3))
subdivide_cluster( base_cl, '1', subdivision1)-> hcl
subdivide_cluster( hcl, '3', subdivision2)-> hcl
get_max_lvl(hcl)
}
\seealso{
\code{\link[=HCCSim_clustering_list]{HCCSim_clustering_list()}} , \code{\link[=subdivide_cluster]{subdivide_cluster()}} for how input data should look like
}
