% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR.R
\name{get_partition_at_g}
\alias{get_partition_at_g}
\title{Extract \code{clustering_list} describing division into the clusters at specified lvl of hierarchy of \code{hclustering_list}}
\usage{
get_partition_at_g(hclustering_list, g)
}
\arguments{
\item{hclustering_list}{A list of clusters and sub-clusters indexed by strings of integer tuples (like '1','2','1,1','1,2','2,1','2,2' etc.). Each cluster is a vector of integers denoting nodes belonging to that cluster.}

\item{g}{An integer specifying lvl of hierarchy (number of subdivision step) to extract from \code{hclustering_list}}
}
\value{
A \code{clustering_list} describing division into the clusters at lvl \code{g} of hierarchy of \code{hclustering_list}
}
\description{
Extract \code{clustering_list} describing division into the clusters at specified lvl of hierarchy of \code{hclustering_list}
}
