% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_utils.R
\name{HCCSim_clustering_list}
\alias{HCCSim_clustering_list}
\title{Initialize clustering_list object}
\usage{
HCCSim_clustering_list(
  x = list(),
  domain_size = length(unlist(x)),
  params = list()
)
}
\arguments{
\item{x}{A list of clusters. Each element is a vector of indexes of nodes that belong to each cluster. Names denote cluster labels. Whatever the names are in the list, they get converted to strings of consecutive integers starting from 1. List can exclude some nodes of originally clustered network, which makes them clusterless.}

\item{domain_size}{Total number of nodes in the network of which node indexes have been placed in each cluster.}

\item{params}{An arbitrary \code{list} that will be stored in an attribute of resulting \code{clustering_list}.}
}
\value{
A \code{HCCSim_clustering_list} object.
}
\description{
\code{clustering_list} is a list of named elements. Each element represents one proper cluster of nodes and contains integers denoting nodes belonging to that cluster.
Name of the element is the cluster label (integer converted to string, e.g. for two proper clusters, there are two elements '1' and '2').
Note that info on clusterless, "noise" nodes is deduced by taking into the account which integers are not contained inside proper clusters.
This requires using attribute \code{'domain_size'} to store the total number of nodes in the network which was clustered.
It posesses additional attribute \code{'params'} which stores optional additional parameters of the clustering algorithm used.
A more friendly function to initialize such object is \code{clustering_vector2list} which accepts as an input a more typical in R representation of clusters, membership vector and converts it to \code{clustering_list}.
}
\examples{
data(brca_clusters)
labels<- sort( unique(brca_clusters[brca_clusters!=0]))
input_list<-list()
for (i in labels)
   input_list[[i]]<- which(brca_clusters==i)
HCCSim_clustering_list(x= input_list, domain_size=length(brca_clusters),  params=list(algorithm='mcl', inflation=2)) 
}
\seealso{
\code{\link[=HCCSim_clustering_vector]{HCCSim_clustering_vector()}} and \code{\link[=clustering_list2vector]{clustering_list2vector()}}
}
