% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_utils.R
\name{HCCSim_clustering_vector}
\alias{HCCSim_clustering_vector}
\title{Initialize clustering_vector object}
\usage{
HCCSim_clustering_vector(x = integer(), params = list())
}
\arguments{
\item{x}{A vector of nonnegative integer labels. Gets coerced to \code{integer} type.}

\item{params}{An arbitrary \code{list} that will be stored in an attribute of resulting \code{clustering_vector}.}
}
\value{
A \code{HCCSim_clustering_vector} object.
}
\description{
\code{clustering_vector} is an integer vector for which \code{clustering_vector[i]} gives a label of the group to which node \code{i} belongs.
It posesses additional attribute \code{'params'} which stores optional additional parameters of the clustering algorithm used.
Labels of proper clusters start from 1. Zero label denotes the "noise" subset of nodes not belonging to any proper cluster.
}
\examples{
data(brca_clusters)
brca_clusters<- HCCSim_clustering_vector(brca_clusters, params=list(algorithm='mcl', inflation=2))
}
\seealso{
\code{\link[=HCCSim_clustering_list]{HCCSim_clustering_list()}} and \code{\link[=clustering_vector2list]{clustering_vector2list()}}
}
