% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR.R
\name{initial_clusterNreconstruct}
\alias{initial_clusterNreconstruct}
\title{Divide variables dataset based on input clustering and reconstruct clusters using at most k principal components}
\usage{
initial_clusterNreconstruct(
  X,
  X_variances,
  clustering_vector = NULL,
  clustering_list = NULL,
  min_cl_size = min(30, ncol(X)),
  f_E = 0.6,
  k = 5
)
}
\arguments{
\item{X}{A matrix containing dataset where columns are variables and rows are samples.}

\item{X_variances}{Precomputed variances of \code{X}.}

\item{clustering_vector}{A vector encoding cluster membership according to convention outlined in \code{HCCsim_clustering_vector} class. Either this or \code{clustering_list} argument shall be non null.}

\item{clustering_list}{A list encoding cluster membership according to convention outlined in \code{HCCsim_clustering_list} class. Either this or \code{clustering_list} argument shall be non null.}

\item{min_cl_size}{If given, nodes in clusters of size less than \code{min_cl_size} get classified as clusterless nodes.}

\item{f_E}{A target fraction of variance to explain. Gets recalculated to match the corresponding fraction of total variance of proper clusters (without clusterless variables).}

\item{k}{Maximum number of principal components to use for reconstructing each cluster.}
}
\value{
A \code{list} with following elements:
\itemize{
\item \code{clustering_list} A \code{clustering_list} representing input clustering.
\item \code{f} Recalculated \code{f_E} giving the fraction of variance of proper clusters to explain (excluding "noise" clusters (\code{C_0}))
\item \code{cluster_blockPCA} A \code{k}-th order clustering based reconstruction of \code{X}, based on \code{clustering_list} component. It is output of \code{blockwise_PCA_reconstruction} funtion.
}
}
\description{
Divide variables dataset based on input clustering and reconstruct clusters using at most k principal components
}
\examples{
data(brca)
data(brca_clusters)
result<- initial_clusterNreconstruct(X= brca, X_variances=matrixStats::colVars(brca),
			        clustering_vector=brca_clusters)
lapply(result$clustering_list, length)
length(result$cluster_blockPCA)
}
\seealso{
\code{\link[=HCCSim_clustering_list]{HCCSim_clustering_list()}} and \code{\link[=HCCSim_clustering_vector]{HCCSim_clustering_vector()}} for how input data should look like, \code{\link[=blockwise_PCA_reconstruction]{blockwise_PCA_reconstruction()}} and \code{\link[=calculate_f]{calculate_f()}} for individual component functions used by this one
}
