% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{KS_ecdf_distance}
\alias{KS_ecdf_distance}
\title{Kolmogorov-Smirnov distance between sample and reference c.d.f.}
\usage{
KS_ecdf_distance(y, ecdf_x)
}
\arguments{
\item{y}{A input vector of values which \code{ecdf} will be compared with the reference.}

\item{ecdf_x}{A function acting on \code{y}, reference empirical c.d.f. (or theoretical of a known distribution), e.g. produced by base R \code{ecdf()} function.}
}
\value{
A KS statistic computed in the following manner: \code{max(abs( ecdf_y(y) - ecdf_x(y)))}.
}
\description{
Kolmogorov-Smirnov distance between sample and reference c.d.f.
}
\examples{
x<- rnorm(1000)
y<- rnorm(1000,0.05)
KS_dist<- KS_ecdf_distance(x, ecdf(y))
}
\seealso{
\code{\link[=ecdf_distance]{ecdf_distance()}} for more general distance between \code{ecdf}s.
}
