% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{MSE}
\alias{MSE}
\title{Compute mean squared error between two vectors}
\usage{
MSE(x, y)
}
\arguments{
\item{x}{One of vectors between which squared error is to be computed.}

\item{y}{One of vectors between which squared error is to be computed.}
}
\value{
A mean of squared errors between \code{x,y}.
}
\description{
Compute mean squared error between two vectors
}
\examples{
MSE( c(1,2,3), c(3,2,1) )
}
