% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCS.R
\name{PC_generator_matrix}
\alias{PC_generator_matrix}
\title{Extract matrix of principal components of all clusters of HCD}
\usage{
PC_generator_matrix(hcluster_blockPCA)
}
\arguments{
\item{\code{hcluster_blockPCA}}{A list of blockwise PCA reconstruction of clusters and subclusters as obtained and \code{subclusterNreconstruct} function.}
}
\value{
A matrix with all prinicipal components inside each block of \code{hcluster_blockPCA} structure. Column names give tuples of clusters to which each prinicipal component belongs.
}
\description{
Extract matrix of principal components of all clusters of HCD
}
\examples{
data(brca)
data(brca_clusters)
lvl1<- initial_clusterNreconstruct(X= brca, X_variances=matrixStats::colVars(brca),
			        clustering_vector=brca_clusters)
lvl2<- subclusterNreconstruct(X=brca,
				 X_variances= matrixStats::colVars(brca),
				 hclustering_list= lvl1$clustering_list,
                              hcluster_PCA_blocks=lvl1$cluster_blockPCA,
                              clfun2=similarity_based_hclust,
                              clfun2OtherArgs_constant=list(method="complete" ),
			 clfun2OtherArgs_ranges= list(n_group=2:7)) 
PC_generator_matrix( lvl2$hcluster_blockPCA) -> PCmat
 #note: no correlations between base clusters and subclusters.
heatmap( sqrt(abs(cor(PCmat))), Colv=NA, Rowv=NA)
}
\seealso{
\code{\link[=initial_clusterNreconstruct]{initial_clusterNreconstruct()}}, \code{\link[=subclusterNreconstruct]{subclusterNreconstruct()}} for how input data should look like
}
