% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_construction.R
\name{rcorrData}
\alias{rcorrData}
\title{Generate list storing all pairwise Pearson correlation coefficients between all variables in the dataset, and their (adjusted) 2-tailed p-values using \code{Hmisc::rcorr}}
\usage{
rcorrData(D, rcorr.method = "pearson", p.adjust.method = "holm")
}
\arguments{
\item{D}{Must be an object of a subclass of \code{matrix}. Rows store samples and columns - variables. No checks on 0 variance are performed.}

\item{p.adjust.method}{The method of p-value adjustment used. For \code{D} with \code{N} columns, number of tests is \code{N*(N-1)*0.5}. Gets passed to argument \code{p.adjust.method} in a call of \code{p.adjust} function.}

\item{NA.warn}{Logical, if true (default), then a warning will be displayed if \code{D} contains any \code{NA} values.}
}
\value{
A list storing all the important data about correlations in \code{D} with the following fields:
\itemize{
\item \code{N} Number of variables in \code{D}.
\item \code{r} A vector of all correlation coefficients between all pairs of variables in \code{D}.
\item \code{P} A vector of unadjusted p-values of \code{r}.
\item \code{Pa} A vector of adjuste p-values of \code{r}.
}
}
\description{
Function utilizes \code{rcorr} from \code{Hmisc} package. It is less speed and memory efficient than \code{fastPearsonData} but is safer to use if input data is not properly cleaned.
}
\examples{
data(brca)
rcorrData(brca) -> brca_corData
C<- matrix(nrow= brca_corData$N, ncol= brca_corData$N) # restore matrices from vectors in corData like this:
C[lower.tri(C)]= brca_corData$r
C<-t(C)
C[lower.tri(C)]= brca_corData$r
diag(C)=1
}
\seealso{
\code{\link[=fastPearsonData]{fastPearsonData()}} for efficient version of correlation coefficient calculation for clean data, \code{\link[Hmisc:rcorr]{Hmisc::rcorr()}} for correlation function used here,
\code{\link[=similarity_matrix]{similarity_matrix()}} for creating similarity matrix out of output of this function
}
