% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR.R
\name{similarity_based_hclust}
\alias{similarity_based_hclust}
\title{Normalized version of hierarchical clustering suited for use with HCR}
\usage{
similarity_based_hclust(
  similarity_matrix,
  n_group,
  max_sim = 1,
  method = "complete"
)
}
\arguments{
\item{similarity_matrix}{A symmetric matrix of node similarities.}

\item{n_group}{A number of groups in the output clustering.}

\item{max_sim}{Maximum value of similarities in \code{similarity_matrix}.}

\item{method}{A method for cluster to cluster distance calculation in \code{hclust}.}
}
\value{
An integer vector of cluster labels of nodes.
}
\description{
Clustering functions used with HCR must take as an input similarity matrix and return vector of cluster labels. This function adapts \code{hclust} from base R to this convention.
}
