% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{strengths}
\alias{strengths}
\title{Compute strengths of nodes from similarity matrix}
\usage{
strengths(A)
}
\arguments{
\item{A}{A similarity matrix  \code{A[i,j]} is weight of the connection between nodes \code{i,j}.}
}
\value{
A vector of strengths of all nodes, weighted versions of degrees.
}
\description{
Compute strengths of nodes from similarity matrix
}
