% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR.R
\name{subclusterNreconstruct}
\alias{subclusterNreconstruct}
\title{Subdivide clusters in the innermost layer of hierarchical clustering and reconstruct the residuals in the subdivided clusters}
\usage{
subclusterNreconstruct(
  X,
  X_variances,
  hclustering_list,
  hcluster_PCA_blocks,
  k = 5,
  f = 0.6,
  clfun2,
  clfun2OtherArgs_constant = NULL,
  clfun2OtherArgs_ranges = NULL,
  S_power = 2,
  corfun = cor
)
}
\arguments{
\item{X}{A matrix containing dataset where columns are variables and rows are samples.}

\item{X_variances}{Precomputed variances of \code{X}.}

\item{hclustering_list}{A list encoding cluster membership, such as one produced by this function or \code{subdivide_cluster} or an instance of \code{HCCSim_clustering_list}. Clusters at the lowest level will be subdivided and reconstructed.}

\item{hcluster_PCA_blocks}{Corresponding reconstructions of blocks defined by clusters from \code{hclustering_list}. Produced by this function or by \code{initial_clusterNreconstruct} or \code{blockwise_PCA_reconstruction}.}

\item{k}{Maximum number of principal components to use for reconstructing each cluster.}

\item{f}{Target fraction of variance to explain (of the proper clusters).}

\item{clfun2}{Clustering function to group residuals of each innermost cluster of \code{hclustering_list} based on magnitude of their correlations. Must receive similarity matrix as a first argument. Best clustering from several different settings will be picked by maximizing \code{normalized_cl_entropy()} value.}

\item{clfun2OtherArgs_constant}{List of named arguments of \code{clfun2} apart from similarity matrix, which shall be kept constant across different candidate settings to test.}

\item{clfun2OtherArgs_ranges}{A list of named vectors of parameters to test. Each named vector in the list shall have the same length. Names of the vectors should correspond to values of named arguments of \code{clfun2}. At step \code{i}, settings of \code{clfun2} are tested defined by \code{i}-th entry of each named vector in that list. Finally, for each cluster, setting maximizing value of \code{normalized_cl_entropy()} function will be chosen as definitive subdivision to use.}

\item{S_power}{A value of exponent which to use for transforming correlation of residuals to similarity by \code{abs(cor_matrix)^S_power}.}

\item{corfun}{A function to compute correlation of residuals of each cluster. Must accept matrix as its 1st argument (where columns are variables and rows are samples.}
}
\value{
A \code{list} with following elements:
\itemize{
\item \code{hclustering_list} An expanded list representing resulting hierarchical clustering, it contains additional clusters that resulted from subdivision.
\item \code{hcluster_blockPCA} A \code{k}-th order clustering based reconstruction of residuals of reconstructions of previous level (determined by input \code{hcluster_blockPCA}. It is expanded by including reconstructions of subclusters added to expanded \code{hclustering_list}.
}
}
\description{
Subdivide clusters in the innermost layer of hierarchical clustering and reconstruct the residuals in the subdivided clusters
}
\examples{
data(brca)
data(brca_clusters)
lvl1<- initial_clusterNreconstruct(X= brca, X_variances=matrixStats::colVars(brca),
			        clustering_vector=brca_clusters)
lvl2<- subclusterNreconstruct(X=brca,
				 X_variances= matrixStats::colVars(brca),
				 hclustering_list= lvl1$clustering_list,
                              hcluster_PCA_blocks=lvl1$cluster_blockPCA,
                              clfun2=similarity_based_hclust,
                              clfun2OtherArgs_constant=list(method="complete" ),
			 clfun2OtherArgs_ranges= list(n_group=2:7)) 
}
\seealso{
\code{\link[=initial_clusterNreconstruct]{initial_clusterNreconstruct()}} for 1st stage clustering and reconstruction, \code{\link[=subdivide_cluster]{subdivide_cluster()}} or \code{\link[=HCCSim_clustering_list]{HCCSim_clustering_list()}}, \code{\link[=blockwise_PCA_reconstruction]{blockwise_PCA_reconstruction()}} for how input clustering and clusterwise PCA blocks should look like, \code{\link[=similarity_based_hclust]{similarity_based_hclust()}} for how a function suitable for \code{clfun2} argument should look like, \code{\link[=corfast]{corfast()}} for efficient correlation calculation that can be used fo \code{corfun}
}
