% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brca_clusters_vector.R
\docType{data}
\name{brca_clusters}
\alias{brca_clusters}
\title{Clusters of genes of BRCA dataset}
\format{
A vector of integers encoding cluster labels of each gene. Zeroes denote clusterless genes.
}
\usage{
data(brca_clusters)
}
\description{
Data description
}
\examples{
data(brca_clusters)
length(brca_clusters)
table(brca_clusters)
}
\references{
Add some
}
\keyword{datasets}
