% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_utils.R
\name{clustering_list2vector}
\alias{clustering_list2vector}
\title{Transform clustering_list to clustering_vector}
\usage{
clustering_list2vector(clustering_list)
}
\arguments{
\item{clustering_list}{A list of named elements representing division of nodes into clusters, assumed to be built with the convention of \code{HCCSim_clustering_list}. Function does not require that argument to be object of the mentioned class but assumes it posseses the same attributes as such object and is structured similarly (attribute \code{domain_size} being crucial!).}
}
\value{
An integer vector for which \code{clustering_vector[i]} gives a label of the group to which node \code{i} belongs. An object of class \code{HCCSim_clustering_vector} with attribute \code{params} copied from the input \code{clustering_list}.
}
\description{
Transform a representation of clusters in terms of a list of indexes to the vector of numeric labels. See fifunctions \code{HCCSim_clustering_vector} and \code{HCCSim_clustering_list}.
}
\examples{
data(brca_clusters)
cl_vec<-HCCSim_clustering_vector(brca_clusters, params =list(algorithm='mcl', inflation=2)) 
table(cl_vec)
clustering_vector2list( cl_vec )-> cl_list
lapply(cl_list, length)
clustering_list2vector(cl_list) -> cl_vec2
all( cl_vec == cl_vec2)
}
