% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_utils.R
\name{clustering_vector2list}
\alias{clustering_vector2list}
\title{Transform clustering_vector to clustering_list}
\usage{
clustering_vector2list(clustering_vector)
}
\arguments{
\item{clustering_vector}{An integer vector for which \code{clustering_vector[i]} gives a label of the group to which node \code{i} belongs. Its attribute \code{'params'} is copied onto the resulting \code{clustering_list}. Does not have to be an object of class \code{HCCSim_clustering_vector} but function assumes it sticks to the labeling convention outlined for that kind of representation.}
}
\value{
A list with named elements with one element per proper cluster, where each element is integer vector of indexes of nodes. An object of class \code{HCCSim_clustering_list}.
}
\description{
\code{clustering_list} is a list of named elements. Each element represents one proper cluster of nodes and contains integers denoting nodes belonging to that cluster.
Name of the element is the cluster label (integer converted to string, e.g. for two proper clusters, there are two elements '1' and '2').
Note that info on clusterless, "noise" nodes is deduced by taking into the account which integers are not contained inside proper clusters.
This requires using attribute \code{'domain_size'} to store the total number of nodes in the network which was clustered.
It posesses additional attribute \code{'params'} which stores optional additional parameters of the clustering algorithm used.
\code{clustering_vector} is an integer vector for which \code{clustering_vector[i]} gives a label of the group to which node \code{i} belongs.
It posesses additional attribute \code{'params'} which stores optional additional parameters of the clustering algorithm used.
Labels of proper clusters start from 1. Zero label denotes the "noise" subset of nodes not belonging to any proper cluster.
}
\examples{
data(brca_clusters)
clustering_vector2list( HCCSim_clustering_vector(brca_clusters, params= =list(algorithm='mcl', inflation=2)) )
}
