% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_construction.R
\name{corfast}
\alias{corfast}
\title{Minimal optimized computation of Pearson correlation by matrix product}
\usage{
corfast(D)
}
\arguments{
\item{D}{Must be an object of a subclass of \code{matrix}. Rows store samples and columns - variables. No \code{NA}s allowed.}
}
\value{
A numeric square matrix of Pearson correlation coefficient compute between each pair of varables in \code{D}.
}
\description{
This function will generally perform faster than base R \code{cor}, provided one uses OpenBLAS matrix multiplication backend.
}
\examples{
data(brca)
C<- corfast(brca)
}
\references{
\link{1}  M Flight, Robert. 2023. “Installing OpenBLAS for Self-Compiled R.” November 1, 2023. \url{https://rmflight.github.io/posts/2023-11-01-installing-openblas-for-selfcompiled-r}
}
\seealso{
\code{\link[=corpval]{corpval()}} and \code{\link[=fastPearsonData]{fastPearsonData()}} for p-values, \href{https://rmflight.github.io/posts/2023-11-01-installing-openblas-for-selfcompiled-r}{this link} \link{1} for how to set up OpenBLAS on Linux,
and \href{https://github.com/david-cortes/R-openblas-in-windows}{that one} for instructions for Windows.
}
