% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR.R
\name{get_max_lvl}
\alias{get_max_lvl}
\title{Extract number of levels of hierarchy from \code{hclustering_list}}
\usage{
get_max_lvl(hclustering_list)
}
\arguments{
\item{hclustering_list}{A list of clusters and sub-clusters indexed by strings of integer tuples (like '1','2','1,1','1,2','2,1','2,2' etc.). Each cluster is a vector of integers dentoing nodes belonging to that cluster.}
}
\value{
An integer giving the number of levels of hierarchy in \code{hclustering_list}
}
\description{
Extract number of levels of hierarchy from \code{hclustering_list}
}
