% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR.R
\name{initial_clusterNreconstruct}
\alias{initial_clusterNreconstruct}
\title{Divide variables dataset based on input clustering and reconstruct clusters using at most k principal components}
\usage{
initial_clusterNreconstruct(
  X,
  X_variances,
  clustering_vector,
  clustering_list = NULL,
  min_cl_size = min(30, ncol(X)),
  f_E = 0.6,
  k = 5
)
}
\arguments{
\item{X}{l}

\item{X_variances}{l}

\item{clustering_vector}{l}

\item{clustering_list}{l}

\item{min_cl_size}{l}

\item{f_E}{l}

\item{k}{l}
}
\value{
A \code{list} with following elements:
\itemize{
\item \code{C_base} A \code{clustering_list} representing input clustering.
\item \code{f} Recalculated \code{f_E} giving the fraction of variance of proper clusters to explain (excluding "noise" clusters (\code{C_0}))
\item \code{C_base_blockPCA} A \code{k}-th order clustering based reconstruction of \code{X}, based on \code{C_base}. It is output of \code{blockwisa_PCA_reconstruction} funtion.
}
}
\description{
Divide variables dataset based on input clustering and reconstruct clusters using at most k principal components
}
\examples{
data(brca)
data(brca_clusters)
result<- initial_clusterNreconstruct(X= brca, X_variances=colVars(brca),
			        clustering_vector=brca_clusters)
lapply(result$C_base, length)
length(result$C_base_blockPCA)
}
