% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCS.R
\name{multivariate_normal_cholesky}
\alias{multivariate_normal_cholesky}
\title{Generate samples from multivariate normal distribution with a given covariance}
\usage{
multivariate_normal_cholesky(sigma, n_samples)
}
\arguments{
\item{sigma}{A covariance matrix to use. Must be positive definite.}

\item{n_samples}{Number of samples to generate}
}
\value{
A matrix of generated samples from a mutlivariate normal distribution with covariance \code{sigma}, each column corresponds to each variable and row to each sample.
}
\description{
Generate samples from multivariate normal distribution with a given covariance
}
\examples{
data(brca)
multivariate_normal_cholesky( cov(brca[,1:300] ), 600) -> gen_data
dim(gen_data)
MSE(cov(brca[,1:300]), cov(gen_data) )
}
