% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR.R
\name{normalized_cl_entropy}
\alias{normalized_cl_entropy}
\title{Calculate entropy of a division of a set into groups}
\usage{
normalized_cl_entropy(clustering_list)
}
\arguments{
\item{clustering_list}{A list where elements are atomic vectors describing separate clusters.  Number of elements in each vector is the number of objects inside that cluster.}
}
\value{
A real number giving the entropy of proportions of groups in the input.
}
\description{
Calculate entropy of a division of a set into groups
}
\examples{
base_cl<- HCCSim_clustering_list(list( `1`= c(5,6,7,8), `2`= c(9,4,3), `3`=c(1,2,10,11) ),
				    params=list(base_param=1) )
normalized_cl_entropy(base_cl)
}
