% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCS.R
\name{PC_generator_matrix}
\alias{PC_generator_matrix}
\title{Extract matrix of principal components of all clusters of HCD}
\usage{
PC_generator_matrix(hcluster_blockPCA)
}
\arguments{
\item{\code{hcluster_blockPCA}}{A list of blockwise PCA reconstruction of clusters and subclusters as obtained and \code{subdivideNreconstruct} function.}
}
\value{
A matrix with all prinicipal components inside each block of \code{hcluster_blockPCA} structure. Column names give tuples of clusters to which each prinicipal component belongs.
}
\description{
Extract matrix of principal components of all clusters of HCD
}
