% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR.R
\name{subclusterNreconstruct}
\alias{subclusterNreconstruct}
\title{Subdivide clusters in the innermost layer of hierarchical clustering and reconstruct the residuals in the subdivided clusters}
\usage{
subclusterNreconstruct(
  X,
  X_similarity_matrix,
  X_variances,
  hclustering_list,
  hcluster_PCA_blocks,
  k = 5,
  f = 0.6,
  clfun2,
  clfun2OtherArgs_constant = NULL,
  clfun2OtherArgs_ranges = NULL,
  S_power = 2,
  corfun = cor
)
}
\arguments{
\item{X}{l}

\item{X_similarity_matrix}{l}

\item{X_variances}{l}

\item{hclustering_list}{l}

\item{hcluster_PCA_blocks}{l}

\item{k}{l}

\item{f}{l}

\item{clfun2}{l}

\item{clfun2OtherArgs_constant}{l}

\item{clfun2OtherArgs_ranges}{l}

\item{S_power}{l}

\item{corfun}{l}
}
\value{
A \code{list}
}
\description{
Subdivide clusters in the innermost layer of hierarchical clustering and reconstruct the residuals in the subdivided clusters
}
