% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCR.R
\name{subclusterNreconstruct}
\alias{subclusterNreconstruct}
\title{Subdivide clusters in the innermost layer of hierarchical clustering and reconstruct the residuals in the subdivided clusters}
\usage{
subclusterNreconstruct(
  X,
  X_similarity_matrix,
  X_variances,
  hclustering_list,
  hcluster_PCA_blocks,
  k = 5,
  f = 0.6,
  clfun2,
  clfun2OtherArgs_constant = NULL,
  clfun2OtherArgs_ranges = NULL,
  S_power = 2,
  corfun = cor
)
}
\arguments{
\item{X}{l}

\item{X_similarity_matrix}{l}

\item{X_variances}{l}

\item{hclustering_list}{l}

\item{hcluster_PCA_blocks}{l}

\item{k}{l}

\item{f}{l}

\item{clfun2}{l}

\item{clfun2OtherArgs_constant}{l}

\item{clfun2OtherArgs_ranges}{l}

\item{S_power}{l}

\item{corfun}{l}
}
\value{
A \code{list}
}
\description{
Subdivide clusters in the innermost layer of hierarchical clustering and reconstruct the residuals in the subdivided clusters
}
\examples{
data(brca)
data(brca_clusters)
lvl1<- initial_clusterNreconstruct(X= brca, X_variances=matrixStats::colVars(brca),
			        clustering_vector=brca_clusters)
lvl2<- subclusterNreconstruct(X=brca, X_similarity_matrix=cor(brca)^2, 
				 X_variances= matrixStats::colVars(brca),
				 hclustering_list= lvl1$C_base,
                              hcluster_PCA_blocks=lvl1$C_base_blockPCA,
                              clfun2=similarity_based_hclust,
                              clfun2OtherArgs_constant=list(method="complete" ),
			 clfun2OtherArgs_ranges= list(n_group=2:7)) 
}
